nanoBragg code changes
changes to the code stored in the simtbx/nanoBragg

+++ simtbx/nanoBragg/nanoBragg.h
@@ -455,7 +455,7 @@ class nanoBragg {
     /* member-wise constructor, allowing all members to be initialized in various ways */
     nanoBragg(
         scitbx::vec2<int> detpixels_slowfast, // = 1024, 1024
-        scitbx::vec3<int> Nabc, // 1 1 1
+        scitbx::vec3<double> Nabc, // 1 1 1
         cctbx::uctbx::unit_cell unitcell, // lysozyme
         vec3 misset, // 0 0 0
         vec2 beam_center, // NAN NAN
